+++ hutool-core/src/main/java/cn/hutool/core/collection/CollUtil.java
@@ -1058,7 +1058,7 @@ public class CollUtil {
 	 * @param <T>        集合元素类型
 	 * @param <K>        唯一键类型
 	 * @param collection 集合
-	 * @param override 是否覆盖模式，如果为{@code true}，加入的新值会覆盖相同key的旧值，否则会忽略新加值
+	 * @param override   是否覆盖模式，如果为{@code true}，加入的新值会覆盖相同key的旧值，否则会忽略新加值
 	 * @return {@link ArrayList}
 	 * @since 5.8.0
 	 */
@@ -3055,95 +3055,5 @@ public class CollUtil {
 		}
 
 		return IterUtil.isEqualList(list1, list2);
-	}
-
-	/**
-	 * 泛型集合向上转型。例如将Collection&lt;Integer&gt;转换为Collection&lt;Number&gt;
-	 * @param collection 集合
-	 * @param <T> 泛型
-	 * @return 泛化集合
-	 */
-	@SuppressWarnings("unchecked")
-	public static <T> Collection<T> castup(Collection<? extends T> collection){
-		return (Collection<T>) collection;
-	}
-
-	/**
-	 * 泛型集合向下转型。例如将Collection&lt;Number&gt;转换为Collection&lt;Integer&gt;
-	 * @param collection
-	 * @param <T>
-	 * @return
-	 */
-	@SuppressWarnings("unchecked")
-	public static <T> Collection<T> castdown(Collection<? super T> collection){
-		return (Collection<T>) collection;
-	}
-
-	/**
-	 * 泛型集合向上转型。例如将Set&lt;Integer&gt;转换为Set&lt;Number&gt;
-	 * @param set 集合
-	 * @param <T> 泛型
-	 * @return 泛化集合
-	 */
-	@SuppressWarnings("unchecked")
-	public static <T> Set<T> castup(Set<? extends T> set){
-		return (Set<T>) set;
-	}
-
-	/**
-	 * 泛型集合向下转型。例如将Set&lt;Number&gt;转换为Set&lt;Integer&gt;
-	 * @param set 集合
-	 * @param <T> 泛型子类
-	 * @return 泛化集合
-	 */
-	@SuppressWarnings("unchecked")
-	public static <T> Set<T> castdown(Set<? super T> set){
-		return (Set<T>) set;
-	}
-
-	/**
-	 * 泛型集合向下转型。例如将Map&lt;Integer, Integer&gt;转换为Map&lt;Number,Number&gt;
-	 * @param map 集合
-	 * @param <K> 泛型父类
-	 * @param <V> 泛型父类
-	 * @return 泛化集合
-	 */
-	@SuppressWarnings("unchecked")
-	public static <K,V> Map<K,V> castup(Map<?extends K, ? extends V> map){
-		return (Map<K,V>) map;
-	}
-
-	/**
-	 * 泛型集合向下转型。例如将Map&lt;Number,Number&gt;转换为Map&lt;Integer, Integer&gt;
-	 * @param map 集合
-	 * @param <K> 泛型子类
-	 * @param <V> 泛型子类
-	 * @return 泛化集合
-	 */
-	@SuppressWarnings("unchecked")
-	public static <K,V> Map<K,V> castdown(Map<?super K, ? super V> map){
-		return (Map<K,V>) map;
-	}
-
-	/**
-	 * 泛型接口向上转型。例如将List&lt;Integer&gt;转换为List&lt;Number&gt;
-	 * @param list 集合
-	 * @param <T> 泛型的父类
-	 * @return 泛化集合
-	 */
-	@SuppressWarnings("unchecked")
-	public static <T> List<T> castup(List<? extends T> list){
-		return (List<T>) list;
-	}
-
-	/**
-	 * 泛型集合向下转型。例如将List&lt;Number&gt;转换为List&lt;Integer&gt;
-	 * @param list 集合
-	 * @param <T> 泛型的子类
-	 * @return 泛化集合
-	 */
-	@SuppressWarnings("unchecked")
-	public static <T> List<T> castdown(List<? super T> list){
-		return (List<T>) list;
 	}
 }
